Camera Can Change Angles
+ Added Camera Lock Functino
+ Added .5seconds buffer for camera lock
+ Add Camera position is tied to GameObjects
