--- Assets/Scripts/FollowPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e76de5a2fd772243bc2d365a0794fbc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
